+++ lib/spack/spack/cmd/bootstrap.py
@@ -12,12 +12,12 @@ import spack
 import spack.bootstrap
 import spack.bootstrap.config
 import spack.bootstrap.core
+import spack.cmd.mirror
 import spack.concretize
 import spack.config
 import spack.llnl.util.filesystem
 import spack.llnl.util.tty
 import spack.llnl.util.tty.color
-import spack.mirrors.utils
 import spack.stage
 import spack.util.path
 import spack.util.spack_yaml
@@ -400,7 +400,7 @@ def _mirror(args):
         spack.llnl.util.tty.set_msg_enabled(False)
         spec = spack.concretize.concretize_one(spec_str)
         for node in spec.traverse():
-            spack.mirrors.utils.create(mirror_dir, [node])
+            spack.cmd.mirror.create(mirror_dir, [node])
         spack.llnl.util.tty.set_msg_enabled(True)
 
     if args.binary_packages:
