+++ lib/spack/spack/test/cmd/bootstrap.py
@@ -9,11 +9,11 @@ import pytest
 
 import spack.bootstrap
 import spack.bootstrap.core
+import spack.cmd.mirror
 import spack.concretize
 import spack.config
 import spack.environment as ev
 import spack.main
-import spack.mirrors.utils
 import spack.spec
 from spack.llnl.path import convert_to_posix_path
 
@@ -184,8 +184,8 @@ def test_bootstrap_mirror_metadata(mutable_config, linux_os, monkeypatch, tmp_pa
     `spack bootstrap add`. Here we don't download data, since that would be an
     expensive operation for a unit test.
     """
-    old_create = spack.mirrors.utils.create
-    monkeypatch.setattr(spack.mirrors.utils, "create", lambda p, s: old_create(p, []))
+    old_create = spack.cmd.mirror.create
+    monkeypatch.setattr(spack.cmd.mirror, "create", lambda p, s: old_create(p, []))
     monkeypatch.setattr(spack.concretize, "concretize_one", lambda p: spack.spec.Spec(p))
 
     # Create the mirror in a temporary folder
